F2F-KIWI-1217-RefactoredCICFE
