+++ .gitignore
@@ -15,6 +15,7 @@ yarn-error.log*
 lerna-debug.log*
 
 # Additional ignores
+.env
 .idea
 .vscode
 public
