+++ test/browser/step_definitions/CICDateOfBirthEntryAndInlineError.step.js
@@ -29,3 +29,23 @@ Given(/^the DOB fields are populated with valid values$/, async function () {
 
     expect(await cdPage.isCurrentPage()).to.be.true;
   });
+
+  Given(/^the user clicks the continue button only on the DoBEntryPage$/, async function () {
+    const dobPage = new DateOfBirthPage(await this.page);
+
+    await dobPage.continue();
+
+  });
+
+
+  Then(/^the user sees an inline error message displayed on the DoBEntryPage$/, async function () {
+    const dobPage = new DateOfBirthPage(await this.page);
+
+    expect(await dobPage.isCurrentPage()).to.be.true;
+
+    const inlineError = 'There is a problem';
+
+    const error = await dobPage.checkErrorText();
+      
+    expect(await error).to.equal(inlineError);
+  });
