+++ test/browser/step_definitions/CICNameEntry.step.js
@@ -2,7 +2,7 @@ const { Given, When, Then} = require("@cucumber/cucumber");
 
 const { expect } = require("chai");
 
-const { NameEntryPage, DateOfBirthPage }  = require("../pages");
+const { NameEntryPage, DateOfBirthPage, EuDrivingLicenceDetailsPageValid }  = require("../pages");
 
 const userData = require("../support/cicUserData.json")
 
@@ -34,4 +34,60 @@ Then(
     expect(await dobPage.isCurrentPage()).to.be.true;
 
   }
+);
+
+Given(/^only one mandatory name field has been entered$/, async function () {
+  const nameEntryPage = new NameEntryPage(await this.page);
+
+  await nameEntryPage.enterSurname();
+  await nameEntryPage.enterMiddleName();
+
+}
+);
+
+When(
+/^the user clicks the continue button in the NameEntry screen$/, async function () {
+const nameEntryPage = new NameEntryPage(await this.page);
+
+await nameEntryPage.continue();
+
+});
+
+Then(
+/^the user sees an inline error message displayed in the NameEntry screen$/,
+async function () {
+  const nameEntryPage = new NameEntryPage(await this.page);
+
+  expect(await nameEntryPage.isCurrentPage()).to.be.true;
+
+  const inlineError = 'There is a problem';
+
+  const error = await nameEntryPage.checkErrorText();
+    
+  expect(await error).to.equal(inlineError);
+  
+}
+);
+
+Given(/^the user has navigated to the Name Entry screen$/, async function () {
+  const nameEntryPage = new NameEntryPage(await this.page);
+
+  expect(await nameEntryPage.isCurrentPage()).to.be.true;
+
+}
+);
+
+When(/^the Back link is clicked on the Name Entry screen$/, async function () {
+const nameEntryPage = new NameEntryPage(await this.page);
+
+await nameEntryPage.back();
+
+});
+
+Then(/^the user is navigated back to the screen that they came from$/,async function () {
+const euDLDetailsPage = new EuDrivingLicenceDetailsPageValid(await this.page);
+
+expect(await euDLDetailsPage.isCurrentPage()).to.be.true;
+  
+}
 );
